Rewrite templates to render instantly
`restia.template` now renders a template to nginx immediately instead of
returning a function that does this.
The option to cache templates has also been removed and they are now
always cached (this was already the default).
Templates are also buffered now, so for every call to `restia.template`
only there's only one call to `ngx.write` at the end.