Switch from restia.template() to restia.templates[]
The template function is now deprecated and the prefered way to render
templates is to access them in the templates table.
This table loads entries dynamically as they are needed and keeps them
in memory after that. Excessively large templates can be removed by
simply setting them to nil in this table.

Restia will first try to load (precompiled) .moonhtml.lua files if
available and then fall back to uncompiled .moonhtml files.