Remove template requirement from cosmo configs completely
Remove the broken dependency for the restia/template module from cosmo
templates in the config module.
This would effectively narrow the definition restia has of a template,
if it wasn't for plain text templates, which already only conform to the
narrower definition that a template can be called like a function.