Add __pairs method to bound config directories
This iterates over the files in a directory and is thus not exactly
super performant, but provides a convenient wrapper around lfs.dir.

The metamethod only works in Lua 5.3 (it's also available as
restia.config.pairs for other platforms) and is only available when lfd
could be required.

Configurations are assumed to have no dots in their name, so the file
extensions need to match exactly. This means:

- "foo.yaml" will be read as a yaml file into foo
- "foo.bar.yaml" will not be loaded, because ".bar.yaml" has no loader
- "foo" will be loaded as plain text in foo
- "foo" (directory) will be bound recursively
- ".foo.yaml" will be skipped